requirements.txt
matplotlib==3.1.2
visdom==0.1.8.9
tqdm==4.46.1
scikit_image==0.16.2
opencv_python==4.2.0.32
numpy==1.18.1
scipy==1.5.0
torchnet==0.0.4
Pillow==8.0.1
